Prepare ViewItem for improved deletion handling
